all code written, now for a lot of debugging

+++ controller.py
@@ -6,11 +6,11 @@ from text import Text
 class Controller:
 
     def __init__(self):
-        self.__drives = []
+        self.drives = []
 
     def add_drive(self, drive):
         if type(drive) == Drive:
-            self.__drives.append(drive)
+            self.drives.append(drive)
             return True
         return False
 
@@ -18,7 +18,7 @@ class Controller:
     def get_entity(self, path):
         path = path.split('\\')
         drive_found = False
-        for drive in self.__drives:
+        for drive in self.drives:
             if drive.name == path[0]:
                 entity = drive
                 drive_found = True
@@ -42,7 +42,7 @@ class Controller:
             return
         parent = self.get_entity(parent_path)
         if type(parent) == Text:
-            raise IllegalFileSystemOperation('parent is a text file')
+            raise IllegalFileSystemOperation('parent cannot be a text file')
         if entity_type == Drive and parent_path:
             raise IllegalFileSystemOperation('drive cannot have a parent')
         try:
@@ -55,11 +55,26 @@ class Controller:
     def delete(self, path):
         self.get_entity(path).delete()
 
-    # raises PathNotFound
+    # raises PathNotFound, IllegalFileSystemOperation, PathAlreadyExists
     def move(self, source_path, destination_path):
-        entity = self.get_entity(source_path)
-        entity.delete()
-        # TODO: add entity at destination_path
+        destination_path = destination_path.split('\\')
+        destination_name = destination_path[-1]
+        destination_path = destination_path[:len(destination_path) - 1]
+        destination_parent = self.get_entity('\\'.join(destination_path))
+        if type(destination_parent) == Text:
+            raise IllegalFileSystemOperation('parent cannot be a text file')
+        try:
+            self.get_entity(destination_path)
+            raise PathAlreadyExists
+        except PathNotFound:
+            entity = self.get_entity(source_path)
+            entity.delete()
+            entity.name = destination_name
+            destination_parent.add_child(entity)
 
+    # raises PathNotFound, IllegalFileSystemOperation
     def write_to_file(self, path, content):
-        pass
+        entity = self.get_entity(path)
+        if type(entity) != Text:
+            raise IllegalFileSystemOperation
+        entity.content = content
